Isometric blocks now overlap vertically
This fixes a problem where hidden blocks were sometimes showing through
to the surface.

Adjusted several constants, as well as height and margin calculations.

Also added a check to avoid drawing the top of a block if it will be
covered by an opaque block or one of the same type. This will save time
and also avoid errors rendering columns of translucent blocks.

This will also pave the way for drawing multicoloured blocks.