Merge most isometric and orthographic render functions
This gets rid of a lot of code duplication.

Also got rid of the heightmap rendering functions, finally.

Also fixed a bug in the out-of-bounds check in
world::get_path_from_rel_coords.

Also, the measure_world function was still switching on rotation for no
reason at all, so removed that.