Changed texture struct format; added isometric block side culling
Sides of blocks are no longer rendered if the block in front is opaque.

Now storing shapes inside the texture struct so we can quickly check the
shape ID for a block.

Shapes are currently defined inside the texture source file, but we'll
find a nicer way of doing this.