Improve night rendering, clean up chunk data variables
Chunk data is now stored in a struct to simplify the signatures of some
of the functions in chunk.c

Isometric renders now use blocklight of neighbouring blocks to light the
block sides; previously the blocklight of the above block was used for
all sides, which resulted in lighting errors. Looks much more natural
now.