Refactor region margin functions
There was no need for two separate functions for calculating region
margins in isometric and top-down modes, so managed to simplify.

Also removed the check in isometric mode to see if the region is on the
edge of the diamond, as it doesn't have to be to be closest to the edge
of the image.

Also renamed some variables to be (hopefully) more clear.