Fix lighting bug
Shadows/hilights need to be either replaced with the main colour in
memory and stay in the colour count, or replaced in the pixmap and
removed from the count. Not sure which is faster, so going with the
latter for now.

(Before this commit they were being replaced in memory and removed from
the count, so they were still being used but weren't being darkened
properly at night.)